requirements.txt
streamlit==1.37.1
pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
easyocr==1.7.1
opencv-python-headless==4.10.0.84
deep-translator==1.11.4
xlsxwriter==3.2.0
